+++ packages/common/src/discussions/_internal/DiscussionUiSchemaCreate.ts
@@ -44,14 +44,6 @@ export const buildUiSchema = async (
           helperText: {
             labelKey: `${i18nScope}.fields.prompt.helperText`,
           },
-          messages: [
-            {
-              type: "ERROR",
-              keyword: "required",
-              icon: true,
-              labelKey: `${i18nScope}.fields.prompt.requiredError`,
-            },
-          ],
         },
       },
     ],