+++ packages/common/src/sites/_internal/SiteBusinessRules.ts
@@ -38,6 +38,7 @@ export const SitePermissions = [
   "hub:site:workspace:followers",
   "hub:site:workspace:followers:member",
   "hub:site:workspace:followers:manager",
+  "hub:site:workspace:followers:create",
   "hub:site:workspace:discussion",
   "hub:site:manage",
 ] as const;
@@ -137,6 +138,9 @@ export const SitesPermissionPolicies: IPermissionPolicy[] = [
   },
   {
     permission: "hub:site:workspace:followers",
+    // TODO: refactor once we have an "upsell" UI to try to
+    // get basic users to switch to premium for this feature
+    licenses: ["hub-premium"],
     dependencies: ["hub:site:workspace", "hub:site:edit"],
   },
   {
@@ -160,6 +164,12 @@ export const SitesPermissionPolicies: IPermissionPolicy[] = [
         value: "entity:followersGroupId",
       },
     ],
+  },
+  // permission to create a followers group
+  {
+    permission: "hub:site:workspace:followers:create",
+    dependencies: ["hub:site:workspace:followers", "hub:group:create"],
+    privileges: ["portal:user:addExternalMembersToGroup"],
   },
   {
     permission: "hub:site:workspace:discussion",