+++ packages/common/test/projects/fixtures.ts
@@ -150,6 +150,7 @@ export const PROJECT_HUB_SEARCH_RESULT: IHubSearchResult = {
     thumbnail: "https://thumbnail/mock-thumbnail.png",
     workspaceRelative: "/mock-relative-workspace-url",
   },
+  rawResult: PROJECT_ITEM,
 };
 
 export const PROJECT_DATA = {