+++ packages/common/test/templates/fixtures.ts
@@ -111,6 +111,7 @@ export const TEMPLATE_HUB_SEARCH_RESULT: IHubSearchResult = {
     thumbnail: "https://thumbnail/mock-thumbnail.png",
     workspaceRelative: "/mock-relative-workspace-url",
   },
+  rawResult: TEMPLATE_ITEM,
 };
 
 export const TEMPLATE_DATA = {};