+++ packages/discussions/src/utils/channel-permission.ts
@@ -153,13 +153,12 @@ export class ChannelPermission {
     const orgPermissions = this.permissionsByCategory[AclCategory.ORG] ?? [];
 
     return orgPermissions.some((permission) => {
-      const { key } = permission;
       if (permission.key !== user.orgId) {
         return false;
       }
 
       return (
-        doesPermissionAllowOrgRole(permission, user.role) &&
+        doesPermissionAllowOrgRole(permission, user) &&
         channelActionLookup(action).includes(permission.role)
       );
     });
@@ -276,12 +275,12 @@ function doesPermissionAllowGroupMemberType(
 
 function doesPermissionAllowOrgRole(
   permission: IChannelAclPermission,
-  orgRole: string
+  user: IDiscussionsUser
 ): boolean {
   return (
     permission.category === AclCategory.ORG &&
     (permission.subCategory === AclSubCategory.MEMBER ||
-      (permission.subCategory === "admin" && orgRole === "org_admin"))
+      (permission.subCategory === AclSubCategory.ADMIN && isOrgAdmin(user)))
   );
 }